table_builder moved to details
