+++ cpp/oneapi/dal/algo/kmeans_init/backend/cpu/train_kernel_dense.cpp
@@ -67,7 +67,7 @@ static train_result call_daal_kernel(const context_cpu& ctx,
                                                                           *(par.engine)));
 
     return train_result().set_centroids(
-        homogen_table_builder{}.reset(arr_centroids, cluster_count, column_count).build());
+        dal::detail::homogen_table_builder{}.reset(arr_centroids, cluster_count, column_count).build());
 }
 
 template <typename Float>
