+++ cpp/oneapi/dal/algo/knn/backend/gpu/infer_kernel_brute_force_dpc.cpp
@@ -70,7 +70,7 @@ static infer_result call_daal_kernel(const context_gpu& ctx,
         &daal_parameter));
 
     return infer_result().set_labels(
-        homogen_table_builder{}.reset(arr_labels, row_count, 1).build());
+        dal::detail::homogen_table_builder{}.reset(arr_labels, row_count, 1).build());
 }
 
 template <typename Float>
