+++ cpp/oneapi/dal/algo/linear_kernel/backend/cpu/compute_kernel_dense.cpp
@@ -61,7 +61,7 @@ static compute_result call_daal_kernel(const context_cpu& ctx,
                                                            &daal_parameter);
 
     return compute_result().set_values(
-        homogen_table_builder{}.reset(arr_values, row_count_x, row_count_y).build());
+        dal::detail::homogen_table_builder{}.reset(arr_values, row_count_x, row_count_y).build());
 }
 
 template <typename Float>
