+++ cpp/oneapi/dal/algo/svm/backend/gpu/infer_kernel_dpc.cpp
@@ -101,8 +101,8 @@ static infer_result call_daal_kernel(const context_gpu& ctx,
 
     return infer_result()
         .set_decision_function(
-            homogen_table_builder{}.reset(arr_decision_function, row_count, 1).build())
-        .set_labels(homogen_table_builder{}.reset(arr_label, row_count, 1).build());
+            dal::detail::homogen_table_builder{}.reset(arr_decision_function, row_count, 1).build())
+        .set_labels(dal::detail::homogen_table_builder{}.reset(arr_label, row_count, 1).build());
 }
 
 template <typename Float>
