+++ cpp/oneapi/dal/table/column_accessor_dpc_test.cpp
@@ -95,7 +95,7 @@ TEST(column_accessor_test, can_get_first_column_from_homogen_table_with_subset_o
 TEST(column_accessor_test, can_get_columns_from_homogen_table_builder) {
     sycl::queue q;
 
-    homogen_table_builder b;
+    detail::homogen_table_builder b;
     b.reset(array<float>::zeros(q, 3 * 2), 3, 2);
     {
         column_accessor<double> acc{ b };
